incubator/i5/requirements.txt
anthropic>=0.39.0
elevenlabs>=1.0.0
httpx>=0.27.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
pydub>=0.25.0
feedgen>=1.0.0
arxiv>=2.0.0
